+++ runway/__init__.py
@@ -1,2 +1,2 @@
 """Set package version."""
-__version__ = '0.36.0'
+__version__ = '0.37.0'