+++ runway/commands/__init__.py
@@ -3,6 +3,7 @@ from .deploy import Deploy # noqa
 from .destroy import Destroy # noqa
 from .dismantle import Dismantle # noqa
 from .gitclean import GitClean # noqa
+from .init import Init # noqa
 from .plan import Plan  # noqa
 from .preflight import Preflight  # noqa
 from .test import Test  # noqa